Add first checker idea

+++ cpp_sources/src/simple_source.cpp
@@ -1,4 +1,6 @@
-int Function()
+#include <cstdint>
+
+uint8_t Function()
 {
-    return 0;
+    return 0u;
 }
